requirements.txt
Django==4
django-countries==7.3.2
djangorestframework==3.13.1

psycopg2-binary==2.9.3

django-celery-beat==2.3.0
celery==5.2.7
redis==4.3.4
flower==1.2.0

django-filter==22.1
djangorestframework-simplejwt==4.8.0

djoser==2.1.0
gunicorn==20.1.0
drf-yasg==1.21.3





